System: Add a `<NordcomProvider/>` component to the Nordstar core.
This will in the future be used to deal with things like overlays,
modals, and other components that need to be rendered outside of the
normal flow of the application. It will most likely also deal with the
theming of the components (instead of using a dedicated
`<ThemeProvider/>`) and the `i18n` support.